add test and move things around a bit
