+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/validate/ValidateNoLosingDigitalVoucher.scala
@@ -1,0 +1,21 @@
+package com.gu.sf_contact_merge.validate
+
+import com.gu.sf_contact_merge.TypeConvert._
+import com.gu.sf_contact_merge.getsfcontacts.GetSfAddress
+import com.gu.sf_contact_merge.getsfcontacts.GetSfAddress.IsDigitalVoucherUser
+import com.gu.util.apigateway.ApiGatewayResponse
+import com.gu.util.reader.Types.ApiGatewayOp
+import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
+import com.gu.util.resthttp.LazyClientFailableOp
+
+object ValidateNoLosingDigitalVoucher {
+
+  def apply(losingContacts: List[LazyClientFailableOp[GetSfAddress.IsDigitalVoucherUser]]): ApiGatewayOp[Unit] =
+    losingContacts.toStream.map {
+      _.value.toApiGatewayOp("get SF address").flatMap {
+        case IsDigitalVoucherUser(true) => ReturnWithResponse(ApiGatewayResponse.notFound("failed validation due to a "))
+        case IsDigitalVoucherUser(false) => ContinueProcessing(())
+      }
+    }.find(_.isComplete).getOrElse(ContinueProcessing(()))
+
+}
